refactor(sdk)!: move stream types to the helpers stream submodule
The stream trigger configs, events, IO inputs, auth/results, and update
operations now live only in @iii-dev/helpers/stream and are removed from the
root SDK public surface. Core state operations import the shared update-op types
from the submodule. IStream and StreamRequest/StreamResponse stay in the root SDK
because they reference core-only types. Rust StreamCallRequest/
StreamJoinLeaveCallRequest are renamed StreamChangeEvent/StreamJoinLeaveEvent.
