+++ sdk/packages/python/iii/src/iii/__init__.py
@@ -27,15 +27,7 @@ from .iii_types import (
     TriggerActionVoid,
     TriggerRequest,
 )
-from .stream import (
-    IStream,
-    StreamChangeEvent,
-    StreamChangeEventDetail,
-    StreamContext,
-    StreamJoinLeaveEvent,
-    StreamJoinLeaveTriggerConfig,
-    StreamTriggerConfig,
-)
+from .stream import IStream
 from .triggers import Trigger, TriggerConfig, TriggerHandler, TriggerTypeRef
 from .types import (
     Channel,
@@ -89,12 +81,6 @@ __all__ = [
     "StreamResponse",
     # Stream
     "IStream",
-    "StreamChangeEvent",
-    "StreamChangeEventDetail",
-    "StreamContext",
-    "StreamJoinLeaveEvent",
-    "StreamJoinLeaveTriggerConfig",
-    "StreamTriggerConfig",
     # Format extraction
     "extract_request_format",
     "extract_response_format",
